# WADNRForestHealthTracker.API/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "WADNRForestHealthTracker.API",
    "version": "1.0"
  },
  "paths": {
...
    "/file-resources/{fileResourceGuidAsString}": {
      "get": {
        "tags": [
          "FileResource"
        ],
        "operationId": "DisplayResource_FileResource",
        "parameters": [
          {
            "name": "fileResourceGuidAsString",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/file-resources/GetWithApiKey/{fileResourceGuidAsString}": {
      "get": {
        "tags": [
          "FileResource"
        ],
        "operationId": "DisplayResourceWithApiKey_FileResource",
        "parameters": [
          {
            "name": "fileResourceGuidAsString",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
...
}